LayoutTests/http/tests/misc/script-defer-write-slow-stylesheet-expected.txt

This tests that a slow stylesheet written by a deferred script delays DOMContentLoaded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.body)['background-color'] is "rgb(255, 0, 0)"
DOMContentLoaded
PASS successfullyParsed is true

TEST COMPLETE
